Updated to v1_03_00
This version has the following changes:
Several functions have had their functionality re-worked. The old functions have been saved as <function name>_OLD. THEY ARE DEPRECIATED!!!!
New GetData method has sensible default arguments, but you can specify a timestamp (if you know it!) and whether to send ReadoutRequests and DataRequests before trying to read the data (for testing).
ReadNextDAQPacket and ReadNextDCSPacket return a DataHeaderPacket or DCSReply packet that is the next in line to be processed. They will perform a DMA if they find that the buffer is empty AFTER reading the packet. They also perform a check BEFORE reading to make sure that the pointers have been populated, for initial reading.